Avoid race condition in get_path on some systems
On some systems, the plugin fails to initialize due to a 'TypeError'.
TypeError: argument of type 'NoneType' is not iterable.
The embedded python code initializes G.paths to an empty map;
however, if a key is not found in the map, the code re-initializes
G.paths using G.fsac.get_paths().  On some systems, the first call
to get_paths() returns 'None'.  Checking 'None' for the key raises
the error. Adding a call to G.fsac.get_paths() when fsac is
initialized avoids the race condition.