Add new config options for extra parameters
Add `fsiExtraInteractiveParameters` and `fsiExtraSharedParameters`,
which should both be used in favor of the old `fsiExtraParameters`.

Some fsi parameters are exclusive to the interpreter and have no
compiler equivalent. FSAC will now pass the shared parameters to the
compiler and not the interactive parameters.

Ionide-vim will append all parameters from both options to the execution
of the interpreter (usually `dotnet fsi`).

`fsiExtraParameters` will be deprecated and removed from FSAC, so it is
best to migrate away from that.

See https://github.com/ionide/FsAutoComplete/issues/1210 for more
detail.